added more cases for sad path

+++ app/graphql/types/input/user_recipes/update_type.rb
@@ -2,7 +2,7 @@ module Types
   module Input
     module UserRecipes
       class UpdateType < Types::BaseInputObject
-        argument :id, Float, required: true
+        argument :id, Integer, required: true
         argument :recipe_rating, Float, required: true
       end
     end
